rename internal package to domain
internal is a special package name in Go which is meant to be used for code which you do not want to make usable outside at all.
but since subpackages are being used, it doesn't make sense to have it named as such
