+++ domain/work/jobs.go
@@ -1,7 +1,7 @@
 package work
 
 import (
-	"delivery-slot-checker/internal/apperrors"
+	"delivery-slot-checker/domain/apperrors"
 	"fmt"
 	"io"
 	"os"
